Move the Elasticsearch helpers into the log storage tree
The shared controller utils no longer reach into the Elasticsearch renderers.
