+++ pkg/controller/logstorage/elastic/elastic_controller.go
@@ -45,6 +45,7 @@ import (
 	"github.com/tigera/operator/pkg/common"
 	"github.com/tigera/operator/pkg/controller/certificatemanager"
 	logstoragecommon "github.com/tigera/operator/pkg/controller/logstorage/common"
+	"github.com/tigera/operator/pkg/controller/logstorage/esutils"
 	"github.com/tigera/operator/pkg/controller/logstorage/initializer"
 	"github.com/tigera/operator/pkg/controller/options"
 	"github.com/tigera/operator/pkg/controller/status"
@@ -77,7 +78,7 @@ type ElasticSubController struct {
 	scheme         *runtime.Scheme
 	status         status.StatusManager
 	provider       operatorv1.Provider
-	esCliCreator   utils.ElasticsearchClientCreator
+	esCliCreator   esutils.ElasticsearchClientCreator
 	clusterDomain  string
 	variant        operatorv1.ProductVariant
 	tierWatchReady *utils.ReadyFlag
@@ -99,7 +100,7 @@ func Add(mgr manager.Manager, opts options.ControllerOptions) error {
 	r := &ElasticSubController{
 		client:         mgr.GetClient(),
 		scheme:         mgr.GetScheme(),
-		esCliCreator:   utils.NewElasticClient,
+		esCliCreator:   esutils.NewElasticClient,
 		tierWatchReady: &utils.ReadyFlag{},
 		status:         status.New(mgr.GetClient(), initializer.TigeraStatusLogStorageElastic, opts.KubernetesVersion),
 		clusterDomain:  opts.ClusterDomain,
@@ -410,7 +411,7 @@ func (r *ElasticSubController) Reconcile(ctx context.Context, request reconcile.
 		esAdminUserSecret = rsecret.CopyToNamespace(common.OperatorNamespace(), esAdminUserSecret)[0]
 	}
 
-	esLicenseType, err = utils.GetElasticLicenseType(ctx, r.client, reqLogger)
+	esLicenseType, err = esutils.GetElasticLicenseType(ctx, r.client, reqLogger)
 	if err != nil {
 		// If LicenseConfigMapName is not found, it means ECK operator is not running yet, log the information and proceed
 		if errors.IsNotFound(err) {
@@ -421,7 +422,7 @@ func (r *ElasticSubController) Reconcile(ctx context.Context, request reconcile.
 		}
 	}
 
-	elasticsearch, err := utils.GetElasticsearch(ctx, r.client)
+	elasticsearch, err := esutils.GetElasticsearch(ctx, r.client)
 	if err != nil {
 		r.status.SetDegraded(operatorv1.ResourceReadError, "An error occurred trying to retrieve Elasticsearch", err, reqLogger)
 		return reconcile.Result{}, err
@@ -619,7 +620,7 @@ func (r *ElasticSubController) handleLogStorageFinalizer(ctx context.Context, ls
 	// instances. So, check if those have been deleted before removing the finalizer.
 	if isTerminating(ls) {
 		// The LogStorage instance is terminating. Check whether ES and Kibana CRs exist.
-		elasticsearch, err := utils.GetElasticsearch(ctx, r.client)
+		elasticsearch, err := esutils.GetElasticsearch(ctx, r.client)
 		if err != nil {
 			r.status.SetDegraded(operatorv1.ResourceReadError, "An error occurred trying to retrieve Elasticsearch", err, reqLogger)
 			return err
