+++ pkg/controller/logstorage/esutils/elasticsearch_cloud.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package esutils
 
 import operatorv1 "github.com/tigera/operator/api/v1"
